requirements.txt
appnope==0.1.0
astroid==2.2.5
autopep8==1.4.4
backcall==0.1.0
boto3==1.9.171
botocore==1.12.171
decorator==4.4.0
Django==2.2.2
django-cors-headers==3.0.2
django-filter==2.1.0
django-guardian==2.0.0
django-rest-framework==0.1.0
django-storages==1.7.1
djangorestframework==3.9.4
djangorestframework-simplejwt==4.3.0
docutils==0.14
ipython==7.5.0
ipython-genutils==0.2.0
isort==4.3.20
jedi==0.13.3
jmespath==0.9.4
lazy-object-proxy==1.4.1
mccabe==0.6.1
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
psycopg2==2.8.2
ptyprocess==0.6.0
pycodestyle==2.5.0
Pygments==2.4.2
PyJWT==1.7.1
pylint==2.3.1
pylint-django==2.0.9
pylint-plugin-utils==0.5
python-dateutil==2.8.0
pytz==2019.1
rope==0.14.0
s3transfer==0.2.1
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.2
typed-ast==1.3.5
urllib3==1.25.3
wcwidth==0.1.7
wrapt==1.11.1
